Added connect code to PreTradeForm
